moved the data for the welfare subpages into a subdirectory, for a little more organization
